[#996] Migrate tests for BumpAllocator into iceoryx2-bb-elementary
